refactor(Learningpath_endpoint): ใส่ handle error ให้ครบทุกอัน & ปรับ response สำหรับ POST DELETE PUT

+++ internal/learning-path/service/path_service.go
@@ -15,6 +15,9 @@ func (s *serviceImpl) GetPaths() ([]model.LearningPath, error) {
 }
 
 func (s *serviceImpl) GetPathDetails(id string) (*model.LearningPath, error) {
+	if id == "" {
+		return nil, apperror.NewBadRequest("user_id is required")
+	}
 	path, err := s.pathRepo.GetLearnningPathByID(id)
 	if err != nil {
 		if err == sql.ErrNoRows {
@@ -37,6 +40,16 @@ func (s *serviceImpl) CreatePath(req model.CreatePathRequest) (string, error) {
 }
 
 func (s *serviceImpl) UpdatePath(id string, req model.UpdatePathRequest) error {
+	if id == "" {
+		return apperror.NewBadRequest("user_id is required")
+	}
+	if req.Title == "" &&
+		req.Objective == "" &&
+		req.Description == "" &&
+		req.CoverImgURL == "" &&
+		req.Status == "" {
+		return apperror.NewBadRequest("request body cannot be empty")
+	}
 	if _, err := s.pathRepo.GetLearnningPathByID(id); err != nil {
 		if err == sql.ErrNoRows {
 			return apperror.NewNotFound("cannot update: path id '%s' not found", id)
@@ -51,6 +64,9 @@ func (s *serviceImpl) UpdatePath(id string, req model.UpdatePathRequest) error {
 }
 
 func (s *serviceImpl) DeletePath(id string) error {
+	if id == "" {
+		return apperror.NewBadRequest("path_id is required")
+	}
 	if err := s.pathRepo.DeleteLearnningPath(id); err != nil {
 		return apperror.NewInternal(err)
 	}
@@ -61,6 +77,9 @@ func (s *serviceImpl) StartPath(pathID string, userID string) error {
 	if userID == "" {
 		return apperror.NewBadRequest("user_id is required")
 	}
+	if pathID == "" {
+		return apperror.NewBadRequest("path_ID is required")
+	}
 	if err := s.pathRepo.EnrollLearnningPathUser(pathID, userID); err != nil {
 		return apperror.NewInternal(err)
 	}
@@ -68,6 +87,12 @@ func (s *serviceImpl) StartPath(pathID string, userID string) error {
 }
 
 func (s *serviceImpl) GetEnrollmentStatus(pathID string, userID string) (*model.PathEnroll, error) {
+	if userID == "" {
+		return nil, apperror.NewBadRequest("user_id is required")
+	}
+	if pathID == "" {
+		return nil, apperror.NewBadRequest("path_id is required")
+	}
 	enroll, err := s.pathRepo.GetLearnningPathEnrollmentStatus(pathID, userID)
 	if err != nil {
 		if err == sql.ErrNoRows {
